pyproject.toml
[project]
name = "farehound"
version = "0.1.0"
description = "Personal flight fare monitoring service with AI-powered deal scoring"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.0",
    "anthropic>=0.40",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "telethon>=1.36",
    "apscheduler>=3.10,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
